requirements.txt
Django==2.2
django-filter==21.1
djangorestframework==3.8.0
flake8==4.0.1
mccabe==0.6.1
pycodestyle==2.8.0
pyflakes==2.4.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
